implemented always validate token in frontend
