+++ mobile/src/components/Navigator.tsx
@@ -1,6 +1,4 @@
-import { useNavigation } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import * as SecureStore from "expo-secure-store";
 import React, { useContext } from "react";
 
 import { AuthContext } from "../contexts/AuthContext";
@@ -11,32 +9,8 @@ import LoginScreen from "../screens/LoginScreen";
 
 const Stack = createNativeStackNavigator();
 
-const validateTokenOnLoad = async () => {
-  const navigation = useNavigation();
-  const { validate } = useContext(AuthContext);
-  let userToken;
-
-  try {
-    /* eslint-disable no-unused-vars */
-    userToken = await SecureStore.getItemAsync("token"); // retrieve token
-  } catch (err) {
-    // Restoring token failed
-    navigation.navigate("LoginScreen");
-  }
-
-  // validate token if it exists to check for expiry
-  const validToken = validate();
-
-  // if token is valid & not expired, navigate to home screen
-  if (validToken) navigation.navigate("HomeScreen");
-};
-
 const Navigator = () => {
   const { isLoggedIn } = useContext(AuthContext);
-
-  React.useEffect(() => {
-    validateTokenOnLoad();
-  }, []);
 
   return (
     <Stack.Navigator initialRouteName="Login" screenOptions={{ headerShown: false }}>
