+++ mobile/src/contexts/AuthContext.tsx
@@ -10,7 +10,7 @@ import { User, initialUser, UserContext } from "./UserContext";
 interface AuthState {
   isLoggedIn: boolean;
   login: (username: string, password: string) => void;
-  validate: () => boolean;
+  validate: () => void;
 }
 
 interface Payload extends User {
@@ -22,7 +22,7 @@ interface Payload extends User {
 const initialState: AuthState = {
   isLoggedIn: false,
   login: () => undefined,
-  validate: () => false,
+  validate: () => undefined,
 };
 
 export const AuthContext = createContext<AuthState>(initialState);
@@ -33,76 +33,101 @@ export const AuthProvider: React.FC = ({ children }) => {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const YSC_SERVER_URI = Constants.manifest?.extra?.YSC_SERVER_URI;
 
+  const login = async (username: string, password: string) => {
+    const params = {
+      username,
+      password,
+    };
+    const url = YSC_SERVER_URI + "auth/login?" + new URLSearchParams(params).toString();
+    const res = await fetch(url, {
+      method: "POST",
+    });
+
+    if (res.ok) {
+      const tokenJson = await res.json();
+      const token = JSON.stringify(tokenJson);
+      // store token & user info
+      await SecureStore.setItemAsync("token", token);
+      const decoded: Payload = jwt_decode(token);
+
+      // store user info in User context
+      const newUserState: User = {
+        username: decoded.username,
+        firstName: decoded.firstName,
+        lastName: decoded.lastName,
+        role: decoded.role,
+        email: decoded.email,
+      };
+      setUserState(newUserState);
+      setIsLoggedIn(true);
+      socket.connect();
+      socket.emit("successful login", decoded.username);
+    } else {
+      console.error("Login request was unsuccessful.");
+    }
+  };
+
+  const validate = async () => {
+    const url = YSC_SERVER_URI + "auth/validate";
+    const tokenRes = await SecureStore.getItemAsync("token");
+
+    if (tokenRes) {
+      const { token } = JSON.parse(tokenRes);
+      const res = await fetch(url, {
+        method: "POST",
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      });
+
+      // token is invalid
+      if (res.status !== 200) {
+        // reset user state
+        setUserState(initialUser);
+        setIsLoggedIn(false);
+        console.log("Couldn't validate token.");
+      }
+
+      // token is valid
+      if (res.status === 200) {
+        const decodedValidation: Payload = jwt_decode(`${tokenRes}`);
+        setIsLoggedIn(true);
+        socket.connect();
+        socket.emit("successful login", decodedValidation.username);
+        console.log("Validated token");
+      }
+    }
+  };
+
   const authContextValue = React.useMemo(
     () => ({
-      login: async (username: string, password: string) => {
-        const params = {
-          username,
-          password,
-        };
-        const url = YSC_SERVER_URI + "auth/login?" + new URLSearchParams(params).toString();
-        const res = await fetch(url, {
-          method: "POST",
-        });
-
-        if (res.ok) {
-          const tokenJson = await res.json();
-          const token = JSON.stringify(tokenJson);
-          // store token & user info
-          await SecureStore.setItemAsync("token", token);
-          const decoded: Payload = jwt_decode(token);
-
-          // store user info in User context
-          const newUserState: User = {
-            username: decoded.username,
-            firstName: decoded.firstName,
-            lastName: decoded.lastName,
-            role: decoded.role,
-            email: decoded.email,
-          };
-          setUserState(newUserState);
-          setIsLoggedIn(true);
-          socket.connect();
-          socket.emit("successful login", decoded.username);
-        } else {
-          console.error("Login request was unsuccessful.");
-        }
-      },
-      validate: async () => {
-        const url = YSC_SERVER_URI + "auth/validate";
-        const tokenRes = await SecureStore.getItemAsync("token");
-        const res = await fetch(url, {
-          method: "POST",
-          headers: {
-            Authorization: `Bearer ${tokenRes}`,
-          },
-        });
-
-        // token is invalid
-        if (res.status !== 200) {
-          // reset user state
-          setUserState(initialUser);
-          setIsLoggedIn(false);
-          console.log("Couldn't validate token.");
-          return false;
-        }
-
-        // token is valid
-        if (res.status === 200) {
-          const decodedValidation: Payload = jwt_decode(`${tokenRes}`);
-          setIsLoggedIn(true);
-          socket.connect();
-          socket.emit("successful login", decodedValidation.username);
-          console.log("Validated token");
-          return true;
-        }
-
-        return false;
-      },
+      login,
+      validate,
       isLoggedIn,
     }),
     [isLoggedIn]
   );
+
+  React.useEffect(() => {
+    const validateTokenOnLoad = async () => {
+      let userToken;
+
+      try {
+        /* eslint-disable no-unused-vars */
+        userToken = await SecureStore.getItemAsync("token"); // retrieve token
+      } catch (err) {
+        // Restoring token failed
+        setIsLoggedIn(false);
+        return;
+      }
+
+      // validate token if it exists to check for expiry
+      // function sets isLoggedIn which sends user to login or home screen
+      validate();
+    };
+
+    validateTokenOnLoad();
+  }, []);
 
   return <AuthContext.Provider value={authContextValue}>{children}</AuthContext.Provider>;
 };
